# server/api/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "生产级 DDD 脚手架项目的 API 文档，包含用户认证、事件驱动等核心功能。",
        "title": "Go DDD Scaffold API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.github.com/shenfay/kiqi",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/admin/menus": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin/Menus"
                ],
                "summary": "获取完整菜单树（管理端）",
                "responses": {
                    "200": {
                        "description": "菜单树",
                        "schema": {
                            "$ref": "#/definitions/github_com_shenfay_kiqi_internal_transport_http_response.SuccessResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_shenfay_kiqi_internal_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_shenfay_kiqi_internal_transport_http_response.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin/Menus"
                ],
                "summary": "创建新菜单",
                "parameters": [
                    {
                        "description": "菜单创建数据",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "icon": {
                                    "type": "string"
                                },
                                "key": {
                                    "type": "string"
                                },
                                "label": {
                                    "type": "string"
                                },
                                "parent_id": {
                                    "type": "string"
                                },
                                "path": {
                                    "type": "string"
                                },
                                "permission": {
                                    "type": "string"
                                },
                                "sort_order": {
                                    "type": "integer"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "创建成功",
                        "schema": {
                            "$ref": "#/definitions/github_com_shenfay_kiqi_internal_transport_http_response.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_shenfay_kiqi_internal_transport_http_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_shenfay_kiqi_internal_transport_http_response.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "菜单Key已存在",
                        "schema": {
                            "$ref": "#/definitions/github_com_shenfay_kiqi_internal_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_shenfay_kiqi_internal_transport_http_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/menus/sort": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin/Menus"
                ],
                "summary": "批量更新菜单排序顺序",
                "parameters": [
                    {
                        "description": "排序数据",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "items": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "id": {
                                                "type": "string"
                                            },
                                            "sort_order": {
                                                "type": "integer"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "排序已更新",
                        "schema": {
                            "$ref": "#/definitions/github_com_shenfay_kiqi_internal_transport_http_response.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_shenfay_kiqi_internal_transport_http_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_shenfay_kiqi_internal_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_shenfay_kiqi_internal_transport_http_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/menus/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin/Menus"
                ],
                "summary": "更新菜单信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "菜单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "菜单更新数据",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "icon": {
                                    "type": "string"
                                },
                                "label": {
                                    "type": "string"
                                },
                                "path": {
                                    "type": "string"
                                },
                                "permission": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "$ref": "#/definitions/github_com_shenfay_kiqi_internal_transport_http_response.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_shenfay_kiqi_internal_transport_http_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_shenfay_kiqi_internal_transport_http_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "菜单不存在",
                        "schema": {
                            "$ref": "#/definitions/github_com_shenfay_kiqi_internal_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_shenfay_kiqi_internal_transport_http_response.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin/Menus"
                ],
                "summary": "删除菜单",
                "parameters": [
                    {
                        "type": "string",
                        "description": "菜单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "$ref": "#/definitions/github_com_shenfay_kiqi_internal_transport_http_response.SuccessResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_shenfay_kiqi_internal_transport_http_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "菜单不存在",
                        "schema": {
                            "$ref": "#/definitions/github_com_shenfay_kiqi_internal_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_shenfay_kiqi_internal_transport_http_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/menus/{id}/status": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin/Menus"
                ],
                "summary": "切换菜单启用/禁用状态",
                "parameters": [
                    {
                        "type": "string",
                        "description": "菜单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "状态已切换",
                        "schema": {
                            "$ref": "#/definitions/github_com_shenfay_kiqi_internal_transport_http_response.SuccessResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_shenfay_kiqi_internal_transport_http_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "菜单不存在",
                        "schema": {
                            "$ref": "#/definitions/github_com_shenfay_kiqi_internal_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_shenfay_kiqi_internal_transport_http_response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/auth/menus": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin/Menus"
                ],
                "summary": "获取当前登录用户可见的菜单树",
                "responses": {
                    "200": {
                        "description": "菜单树",
                        "schema": {
                            "$ref": "#/definitions/github_com_shenfay_kiqi_internal_transport_http_response.SuccessResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_shenfay_kiqi_internal_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_shenfay_kiqi_internal_transport_http_response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_shenfay_kiqi_internal_transport_http_response.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {},
                "message": {
                    "type": "string"
                },
                "timestamp": {
                    "description": "响应时间（RFC3339 格式）",
                    "type": "string"
                },
                "trace_id": {
                    "description": "链路追踪 ID（用于日志关联和分布式追踪）",
                    "type": "string"
                }
            }
        },
        "github_com_shenfay_kiqi_internal_transport_http_response.SuccessResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "timestamp": {
                    "description": "响应时间（RFC3339 格式）",
                    "type": "string"
                },
                "trace_id": {
                    "description": "链路追踪 ID（用于日志关联和分布式追踪）",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "使用 JWT Token，格式：Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}